fix(client): codex round 2 — collision keys, fleetActions, typecheck
Per-host computeTerminalKeys without TerminalId map collisions; move fleet
→ palette projection into palette/fleetActions; exact deep-link only; always
IntentMarkdown for intents; rootHidden only on empty browse; perHostName +
readWithFallback for cold activity window; shell budget 2.
